[add] Moved reply and parsing methods from module.c to specific C/header files
